chore(deps): upgrade dependencies and migrate to streamable HTTP transport
- upgraded rmcp from 0.9.1 to 0.11.0
- replaced SSE transport with streamable HTTP in MCP servers
- updated tree-sitter from 0.25.10 to 0.26.2
- fixed Node API calls to use u32 instead of usize in parsers
- updated fastembed from 5.3.1 to 5.4.0
- updated axum-server from 0.7.3 to 0.8.0
- updated tower-http from 0.6.7 to 0.6.8
- updated git2 from 0.20.2 to 0.20.3
- updated criterion from 0.8.0 to 0.8.1
- updated testcontainers from 0.25.2 to 0.26.0
- changed MCP endpoint from /mcp/sse to /mcp
- changed MCP client config type from sse to http
- removed transport-sse-server feature
- fixed clippy warnings in Swift test files
- updated HTTP server documentation

Breaking: MCP clients must update config from type:sse to type:http and URL from /mcp/sse to /mcp

+++ src/parsing/c/parser.rs
@@ -74,7 +74,7 @@ impl CParser {
 
         // Recursively process children
         for i in 0..node.child_count() {
-            if let Some(child) = node.child(i) {
+            if let Some(child) = node.child(i as u32) {
                 Self::extract_imports_from_node(child, code, file_id, imports);
             }
         }
@@ -802,7 +802,7 @@ impl CParser {
 
         // Process children
         for i in 0..node.child_count() {
-            if let Some(child) = node.child(i) {
+            if let Some(child) = node.child(i as u32) {
                 Self::extract_calls_from_node(child, code, calls);
             }
         }
@@ -832,7 +832,7 @@ impl CParser {
 
         // Process children
         for i in 0..node.child_count() {
-            if let Some(child) = node.child(i) {
+            if let Some(child) = node.child(i as u32) {
                 Self::find_calls_in_node(child, code, calls);
             }
         }
@@ -857,7 +857,7 @@ impl CParser {
 
         // Process children
         for i in 0..node.child_count() {
-            if let Some(child) = node.child(i) {
+            if let Some(child) = node.child(i as u32) {
                 Self::find_uses_in_node(child, code, uses);
             }
         }
@@ -910,7 +910,7 @@ impl CParser {
 
         // Process children
         for i in 0..node.child_count() {
-            if let Some(child) = node.child(i) {
+            if let Some(child) = node.child(i as u32) {
                 Self::find_defines_in_node(child, code, defines);
             }
         }
