+++ src/parsing/python/parser.rs
@@ -2824,7 +2824,7 @@ class Outer:
             }
 
             for i in 0..node.child_count() {
-                if let Some(child) = node.child(i) {
+                if let Some(child) = node.child(i as u32) {
                     print_node_with_fields(child, code, depth + 1);
                 }
             }
@@ -2878,7 +2878,7 @@ class Outer:
             }
 
             for i in 0..node.child_count() {
-                if let Some(child) = node.child(i) {
+                if let Some(child) = node.child(i as u32) {
                     print_node_with_fields(child, code, depth + 1);
                 }
             }
